Start prometheus proxy nginx in process 1
**Problem:**
When we start nginx in our start-up script, the nginx process would
become the child process of start-up script process and not process 1.
In this case, the kill signal from kubelet/docker will be sent to
start-up script instead of nginx so the nginx process won't stop after
kill.

**Solution:**
Change the proxy command and let nginx start at process 1.