Remove namespace template
Removed namespace template. Prior, the cattle-system namespace
would be uninstalled on app uninstall causing all resources
contained in the cattle-system namespace to be uninstalled. It
is not necessary for the namespace to be a part of the chart.
The cattle-system namespace will likely exist prior to an
attempt to deploy the app. If not, an error will be returned
in rancher and the app will be deployed once the namespace is
available.