Update wmi_exporter to v0.0.5
Related Issue: https://github.com/rancher/rancher/issues/27911

Why is it related?

In Feb 11, 2020 Windows rolled out a security update that changed the interface between user mode and kernel mode, so difficult-to-debug issues related to starting the container process can show up if the Windows Server container image and the Windows container host are out of sync with the patch.

This is most likely the case with the AWS AMI that experienced the issue noted on the ticket, so the fix from our end is to bump the version of wmi_exporter to one that was built after the patch was released.

Link to Announcement: https://support.microsoft.com/en-us/topic/you-might-encounter-issues-when-using-windows-server-containers-with-the-february-11-2020-security-update-release-b9a8fcae-950d-7a0b-ac7c-cb6b294cb809

Signed-off-by: Arvind Iyengar <arvind.iyengar@rancher.com>